Fix reporting overflow exception (#372)
- Fix reporting overflow exception

+++ src/GRA.Controllers/ViewModel/MissionControl/Reporting/ReportingViewModel.cs
@@ -7,9 +7,9 @@ namespace GRA.Controllers.ViewModel.MissionControl.Reporting
         public GRA.Domain.Model.StatusSummary Request { get; set; }
         public IEnumerable<GRA.Domain.Model.StatusSummary> StatusSummaries { get; set; }
         public int TotalUsers { get; set; }
-        public int TotalBooksRead { get; set; }
-        public int TotalChallengesCompleted { get; set; }
-        public int TotalPointsEarned { get; set; }
-        public int TotalBadgesEarned { get; set; }
+        public long TotalBooksRead { get; set; }
+        public long TotalChallengesCompleted { get; set; }
+        public long TotalPointsEarned { get; set; }
+        public long TotalBadgesEarned { get; set; }
     }
 }
