+++ src/GRA.Domain.Repository/IUserLogRepository.cs
@@ -9,9 +9,9 @@ namespace GRA.Domain.Repository
     {
         Task<IEnumerable<Model.UserLog>> PageHistoryAsync(int userId, int skip, int take);
         Task<int> GetHistoryItemCountAsync(int userId);
-        Task<int> CompletedChallengeCountAsync(StatusSummary request);
-        Task<int> PointsEarnedTotalAsync(StatusSummary request);
-        Task<Dictionary<string, int>> ActivityEarningsTotalAsync(StatusSummary request);
-        Task<int> EarnedBadgeCountAsync(StatusSummary request);
+        Task<long> CompletedChallengeCountAsync(StatusSummary request);
+        Task<long> PointsEarnedTotalAsync(StatusSummary request);
+        Task<Dictionary<string, long>> ActivityEarningsTotalAsync(StatusSummary request);
+        Task<long> EarnedBadgeCountAsync(StatusSummary request);
     }
 }
